refactor(ri): colocate component tests alongside implementations
Move 7 test files from src/__tests__/ to their respective component
directories and update import paths accordingly.
